# docs/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "description": "Fluxend is backend as-a-service platform that allows you to build, deploy, and scale applications without managing infrastructure.",
        "title": "Fluxend API",
        "contact": {
            "name": "API Support",
            "url": "http://github.com/fluxend/fluxend",
            "email": "hello@fluxend.app"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/users": {
            "post": {
                "description": "Add a new user with a name, email, and password",
                "tags": [
                    "Users"
                ],
                "summary": "Create user",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/user.CreateRequest"
                            }
                        }
                    },
                    "description": "User details",
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "User created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/response.Response"
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "content": {
                                                    "$ref": "#/components/schemas/user.Response"
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "422": {
                        "description": "Unprocessable entity"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/users/login": {
            "post": {
                "description": "Authenticate a user and return a JWT token",
                "tags": [
                    "Users"
                ],
                "summary": "Authenticate user",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/user.LoginRequest"
                            }
                        }
                    },
                    "description": "Login request",
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "User details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/response.Response"
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "content": {
                                                    "$ref": "#/components/schemas/user.Response"
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/users/logout": {
            "post": {
                "description": "Invalidate the JWT token to log out a user",
                "tags": [
                    "Users"
                ],
                "summary": "Logout user",
                "parameters": [
                    {
                        "description": "Bearer Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User logged out",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response.Response"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/users/{userUUID}": {
            "get": {
                "description": "Get details of a specific user",
                "tags": [
                    "Users"
                ],
                "summary": "Retrieve user",
                "parameters": [
                    {
                        "description": "Bearer Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "User UUID",
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/response.Response"
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "content": {
                                                    "$ref": "#/components/schemas/user.Response"
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "put": {
                "description": "Update a user's details such as name, email, and password",
                "tags": [
                    "Users"
                ],
                "summary": "Update user",
                "parameters": [
                    {
                        "description": "Bearer Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "User UUID",
                        "name": "userUUID",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/user.UpdateRequest"
                            }
                        }
                    },
                    "description": "User details",
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "User updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/response.Response"
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "content": {
                                                    "$ref": "#/components/schemas/user.Response"
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Unprocessable entity"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        }
    },
    "servers": [
        {
            "url": "//api.fluxend.app"
        }
    ],
    "components": {
...
        "schemas": {
...
            "response.Response": {
                "type": "object",
                "properties": {
                    "content": {},
                    "errors": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "success": {
                        "type": "boolean"
                    }
                }
            },
            "user.CreateRequest": {
                "type": "object",
                "properties": {
                    "bio": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "projectUUID": {
                        "type": "string"
                    },
                    "username": {
                        "type": "string"
                    }
                }
            },
            "user.LoginRequest": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "projectUUID": {
                        "type": "string"
                    }
                }
            },
            "user.Response": {
                "type": "object",
                "properties": {
                    "bio": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "roleId": {
                        "type": "integer"
                    },
                    "status": {
                        "type": "string"
                    },
                    "updatedAt": {
                        "type": "string"
                    },
                    "username": {
                        "type": "string"
                    },
                    "uuid": {
                        "type": "string"
                    }
                }
            },
            "user.UpdateRequest": {
                "type": "object",
                "properties": {
                    "bio": {
                        "type": "string"
                    },
                    "projectUUID": {
                        "type": "string"
                    }
                }
            }
        }
    }
}